requirements.txt
flask
flask-cors
faiss-cpu
sentence-transformers
requests
numpy
pandas
openpyxl
pymupdf
